+++ HISTORY.md
@@ -2,3 +2,5 @@ History
 =======
 
 ## Current
+
+* Implement `--action=read`.